requirements.txt
pyautogui==0.9.52
Pillow==8.2.0
keyboard==0.13.5